+++ paper_uploads/forms/dialogs/base.py
@@ -3,7 +3,7 @@ from django.utils.translation import gettext_lazy as _
 from paper_admin.admin.renderers import PaperFormRenderer
 
 
-class UploadedFileBaseForm(forms.ModelForm):
+class ChangeFileResourceDialogBase(forms.ModelForm):
     default_renderer = PaperFormRenderer
     new_name = forms.CharField(required=True, label=_("File name"), max_length=255,)