+++ paper_uploads/forms/dialogs/collection.py
@@ -1,16 +1,16 @@
 from django import forms
 
 from ...models.collection import FileItem, ImageItem
-from .base import UploadedFileBaseForm
+from .base import ChangeFileResourceDialogBase
 
 
-class FileItemDialog(UploadedFileBaseForm):
+class ChangeFileItemDialog(ChangeFileResourceDialogBase):
     class Meta:
         model = FileItem
         fields = ("new_name", "display_name")
 
 
-class ImageItemDialog(UploadedFileBaseForm):
+class ChangeImageItemDialog(ChangeFileResourceDialogBase):
     class Meta:
         model = ImageItem
         fields = ("new_name", "title", "description")